+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change log
 
 - [#583](https://github.com/mobilityhouse/ocpp/issues/583) OCPP v1.6/v2.0.1 deprecate dataclasses from calls and call results with the suffix 'Payload'
+- [#593](https://github.com/mobilityhouse/ocpp/issues/593) Update tests to use Call and CallResult without the suffix Payload
 
 ## 0.26.0 (2024-01-17)